pull identity adapter, routing_rules, endpoint, and peer_spec controllers from upstream Talos
Replace hand-written controllers with upstream Talos source pulled via
genrule from the pinned http_archive. Each upstream file is either copied
verbatim (identity adapter, routing_rules, endpoint) or minimally
transformed via awk (peer_spec: cluster.Identity→kubespan.Identity,
NodeID→PublicKey, endpoint filtering injection).

- peerstate/: add identity adapter genrule (verbatim copy)
- identity/: remove GenerateKey/UpdateAddress, delegate to upstream adapter
- routing/: replace routing.go with upstream routing_rules genrule + new
  kubespand-only routes.go for InstallRoutes
- endpoint/: replace hand-written controller with upstream genrule
- peerspec/: new package with awk-transformed upstream peer_spec.go and
  kubespand-specific endpoint filtering helpers
- controller_manager.go: add TODO about considering upstream pull
