+++ cluster/kubespan_agent/identity/identity.go
@@ -1,6 +1,7 @@
 // Package identity handles KubeSpan node identity (WireGuard keypair + derived addresses).
 //
-// Adapter functions match Talos internal/app/machined/pkg/adapters/kubespan/identity.go.
+// Key generation and address derivation are in the upstream adapter
+// (peerstate package, pulled from Talos adapters/kubespan/identity.go).
 // LoadOrCreate and DetectMAC are kubespand-only (Talos uses STATE partition and COSI HardwareAddr).
 package identity
 
@@ -8,64 +9,14 @@ import (
 	"errors"
 	"fmt"
 	"net"
-	"net/netip"
 	"os"
 	"path/filepath"
 
-	"github.com/mdlayher/netx/eui64"
-	"github.com/siderolabs/gen/value"
 	"github.com/siderolabs/talos/pkg/machinery/resources/kubespan"
-	"github.com/siderolabs/talos/pkg/machinery/resources/network"
-	"go4.org/netipx"
-	"golang.zx2c4.com/wireguard/wgctrl/wgtypes"
 	"gopkg.in/yaml.v3"
+
+	kubespanadapter "github.com/agentydragon/ducktape/cluster/kubespan_agent/peerstate"
 )
-
-// GenerateKey generates a new WireGuard key pair.
-// Ref: talos/internal/app/machined/pkg/adapters/kubespan/identity.go (GenerateKey)
-func GenerateKey(spec *kubespan.IdentitySpec) error {
-	key, err := wgtypes.GeneratePrivateKey()
-	if err != nil {
-		return err
-	}
-
-	spec.PrivateKey = key.String()
-	spec.PublicKey = key.PublicKey().String()
-
-	return nil
-}
-
-// UpdateAddress re-calculates node address based on cluster ID and MAC.
-// Ref: talos/internal/app/machined/pkg/adapters/kubespan/identity.go (UpdateAddress)
-func UpdateAddress(spec *kubespan.IdentitySpec, clusterID string, mac net.HardwareAddr) error {
-	spec.Subnet = network.ULAPrefix(clusterID, network.ULAKubeSpan)
-
-	var err error
-
-	spec.Address, err = wgEUI64(spec.Subnet, mac)
-
-	return err
-}
-
-// wgEUI64 computes an EUI-64 address within the given prefix from a MAC address.
-// Ref: talos/internal/app/machined/pkg/adapters/kubespan/identity.go (wgEUI64)
-func wgEUI64(prefix netip.Prefix, mac net.HardwareAddr) (out netip.Prefix, err error) {
-	if value.IsZero(prefix) {
-		return out, errors.New("cannot calculate IP from zero prefix")
-	}
-
-	stdIP, err := eui64.ParseMAC(netipx.PrefixIPNet(prefix).IP, mac)
-	if err != nil {
-		return out, fmt.Errorf("failed to parse MAC into EUI-64 address: %w", err)
-	}
-
-	ip, ok := netipx.FromStdIP(stdIP)
-	if !ok {
-		return out, fmt.Errorf("failed to parse intermediate standard IP %q", stdIP.String())
-	}
-
-	return netip.PrefixFrom(ip, ip.BitLen()), nil
-}
 
 // LoadOrCreate loads an existing identity from disk, or generates a new one.
 // The identity file uses YAML format matching Talos's kubespan-identity.yaml.
@@ -79,7 +30,7 @@ func LoadOrCreate(path string, clusterID string) (*kubespan.IdentitySpec, error)
 	}
 
 	spec = &kubespan.IdentitySpec{}
-	if err := GenerateKey(spec); err != nil {
+	if err := kubespanadapter.IdentitySpec(spec).GenerateKey(); err != nil {
 		return nil, fmt.Errorf("generating WireGuard key: %w", err)
 	}
