+++ cluster/kubespan_agent/peerspec/BUILD.bazel
@@ -1,31 +1,36 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
-# Upstream Talos endpoint controller pulled at a pinned tag.
+# Upstream Talos peer spec controller pulled at a pinned tag and transformed.
 # See MODULE.bazel http_archive "talos_internal" for the pinned version.
 
 _TALOS_CONTROLLERS = "@talos_internal//:internal/app/machined/pkg/controllers/kubespan/"
 
 genrule(
-    name = "upstream_endpoint",
-    srcs = [_TALOS_CONTROLLERS + "endpoint.go"],
-    outs = ["endpoint_upstream.go"],
-    cmd = "cp $< $@",
+    name = "upstream_peer_spec",
+    srcs = [_TALOS_CONTROLLERS + "peer_spec.go"],
+    outs = ["peer_spec_upstream.go"],
+    cmd = "awk -f $(location :transform_peer_spec.awk) $< > $@",
+    tools = [":transform_peer_spec.awk"],
 )
 
 go_library(
-    name = "endpoint",
-    srcs = [":upstream_endpoint"],
-    importpath = "github.com/agentydragon/ducktape/cluster/kubespan_agent/endpoint",
+    name = "peerspec",
+    srcs = [
+        "peer_spec_filters.go",
+        ":upstream_peer_spec",
+    ],
+    importpath = "github.com/agentydragon/ducktape/cluster/kubespan_agent/peerspec",
     visibility = ["//cluster/kubespan_agent:__subpackages__"],
     deps = [
         "@com_github_cosi_project_runtime//pkg/controller",
+        "@com_github_cosi_project_runtime//pkg/resource",
         "@com_github_cosi_project_runtime//pkg/safe",
         "@com_github_cosi_project_runtime//pkg/state",
         "@com_github_siderolabs_gen//optional",
-        "@com_github_siderolabs_gen//value",
         "@com_github_siderolabs_talos_pkg_machinery//resources/cluster",
         "@com_github_siderolabs_talos_pkg_machinery//resources/config",
         "@com_github_siderolabs_talos_pkg_machinery//resources/kubespan",
+        "@org_go4_netipx//:netipx",
         "@org_uber_go_zap//:zap",
     ],
 )
